[TEST] - Export model DTO from another file
