+++ src/profile/dto/delete-profile.dto.ts
@@ -1,4 +1,4 @@
-import { profileSchema } from '@/profile/dto/profile.dto';
+import { profileSchema } from '@/prisma/dtos.dto';
 import { createZodDtoWithoutDate } from '@/shared/dto-modification/create-zod-dto-without-date';
 
 export const deleteProfileResponseSchema = profileSchema;
