+++ src/profile/dto/find-with-active-chat.dto.ts
@@ -1,5 +1,5 @@
 import { messageSchema } from '@/message/dto/message.dto';
-import { profileSchema } from '@/profile/dto/profile.dto';
+import { profileSchema } from '@/prisma/dtos.dto';
 import { createZodDtoWithoutDate } from '@/shared/dto-modification/create-zod-dto-without-date';
 import { tagSchema } from '@/tag/dto/tag.dto';
 import z from 'zod';
